Modal is now seperate component
